size_t prints cassed ull

+++ src/jet_token.c
@@ -37,17 +37,17 @@ void jet_token_print_list(jet_list* tokens)
         return;
     }
     
-    printf("\ntoken list [%zu]:\n\n", len);
+    printf("\ntoken list [%llu]:\n\n", (unsigned long long)len);
     printf("%-7s %-17s %-10s %s\n\n", "len", "type", "enum-id", "value");        
    
     for(size_t i = 0; i < len; i++)
     {  
         jet_token* token = jet_list_get(tokens, i);
         const char* type_str = jet_token_type_str(token->type);
-        printf("%-7zu %-17s %-10d %.*s\n",i+1, type_str,(int)token->type, (int)token->len, token->source + token->origin);        
+        printf("%-7llu %-17s %-10d %.*s\n",(unsigned long long)(i+1), type_str,(int)token->type, (int)token->len, token->source + token->origin);        
     }
     printf("\n");
-    printf("* generated %zu tokens\n\n", len);
+    printf("* generated %llu tokens\n\n", (unsigned long long)len);
 }
 
 const char* jet_token_type_str(jet_token_type type)
